wot_client/requirements.txt
# WoT Game Client (Windows)
# 用于游戏画面捕获、操作控制和与训练服务器通信

# Game Automation
pywin32>=305      # Windows API
psutil>=5.9.0     # Process management
mss>=9.0.0        # Screen capture
pynput>=1.7.6     # Input control

# Computer Vision (轻量级，仅用于预处理)
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# API Integration (与服务器通信)
requests>=2.31.0
aiohttp>=3.8.0
websockets>=12.0  # WebSocket 客户端

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Development
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0

# OpenCV 及其视频流/截图支持
opencv-contrib-python>=4.8.0     # 带扩展模块的 OpenCV
imageio[ffmpeg]>=2.31.0          # 视频读写，ffmpeg 支持
pyvirtualcam>=0.10.0             # 虚拟摄像头支持（可选）


